investment_agent/requirements.txt
langchain
langchain-groq
langchain-community
langgraph
duckduckgo-search
yfinance
streamlit
python-dotenv
pandas